// sfr-serverless/common/resources/errors/types.ts
export enum ErrorKey {
  RecordNotFound = "record_not_found",
  ErrorUnknown = "error_unknown",
  Permission = "permission",
  ProfileNotFound = "404",
  Unauthorized = "unauthorized",
  IsAdmin = "is_admin",
  AllowForgotPassword = "allow_forgot_password",

  CountCompetitionError = "count_competition_error",
  CompetitionNotRegisterUpload = "competition_not_register_upload",

  UsernameExisted = "username_existed",
  AuthExisted = "auth_existed",
  EmailExisted = "email_existed",
  EmailNotFound = "email_not_found",
  NotFollowYourSelf = "not_follow_your_self",

  DeleteFeed = "delete_feed",
  CountAttachment = "count_attachment",
  FollowBockUser = "follow_block_user",
  ActionUserBlock = "action_user_block",

  CanNotUntagFeed = "cannot_untag_feed",
  CanNotDoActionForComment = "cannot_do_action_for_comment"
}